pyproject.toml
[tool.poetry]
name = "nonebot-plugin-access-control-api"
version = "1.1.0"
description = ""
authors = ["ssttkkl <huang.wen.long@hotmail.com>"]
license = "MIT"
readme = "README.MD"
repository = "https://github.com/bot-ssttkkl/nonebot-plugin-access-control"
packages = [
    { include = "nonebot_plugin_access_control_api", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.1.0"
nonebot-plugin-session = "^0.2.0"
ssttkkl-nonebot-utils = ">=0.1.17"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.275"
isort = "^5.10.1"
black = "^23.1.0"
pre-commit = "^3.1.0"

setuptools = "^68.1.2"
nb-cli = "^1.2.5"

nonebot-adapter-onebot = "*"
nonebot-adapter-kaiheila = "*"
nonebot-adapter-qqguild = "*"
nonebot-adapter-telegram = "*"
nonebot-adapter-console = "*"
nonebot-adapter-feishu = "*"
nonebot-plugin-apscheduler = "^0.3.0"


[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["C901", "E402", "F403", "T201"]

line-length = 160
target-version = "py39"

[tool.ruff.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.nonebot]
plugins = ["nonebot_plugin_access_control_api"]
plugin_dirs = []

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
